Remove : (previous) Collider2 class Implement : Collider class Revise : ColliderManager class for Collider class Implement : TilemapEditorScene and DrawTilemap class
타일맵을 그리기위한 타일맵 툴인 TilemapEditorScene 클래스 추가
저장한 타일맵 데이터를 게임씬에 그리기위한 DrawTilemap 클래스 추가.
타일맵 이미지 추가.
이전에 사용하던 Collider2 클래스가 문제가 있어서, 새로이 Collider 클래스를 만듦.
모든 콜라이더를 관리하는 ColliderManager 클래스가 새로 만든 Collider클래스와 호환되게 코드 수정.
이전에 사용하던 Collider2 클래스와 관련된 코드를 삭제하고, 새로 만든 Collider 클래스의 코드 추가.

+++ AmmoManager.cpp
@@ -1,11 +1,10 @@
 #include "stdafx.h"
 #include "AmmoManager.h"
 #include "Ammo.h"
-#include "Collider2.h"
 
 void AmmoManager::Update()
 {
-	int size = m_vecAmmos.size();
+	size_t size = m_vecAmmos.size();
 
 	for (size_t i = 0; i < size; ++i)
 	{
@@ -26,7 +25,7 @@ void AmmoManager::Update()
 
 void AmmoManager::Render(HDC hdc)
 {
-	int size = m_vecAmmos.size();
+	size_t size = m_vecAmmos.size();
 
 	for (size_t i = 0; i < size; ++i)
 	{
@@ -36,7 +35,7 @@ void AmmoManager::Render(HDC hdc)
 
 void AmmoManager::Release()
 {
-	int size = m_vecAmmos.size();
+	size_t size = m_vecAmmos.size();
 
 	for (size_t i = 0; i < size; ++i)
 	{
